% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\docType{methods}
\name{BICplot}
\alias{BICplot}
\alias{BICplot,numeric,numeric,character-method}
\title{BICplot}
\usage{
BICplot(g, BIC, pdffout)

\S4method{BICplot}{numeric,numeric,character}(g, BIC, pdffout)
}
\arguments{
\item{g}{A numeric vector of cluster sizes.}

\item{BIC}{A numeric vector of BIC values.}

\item{pdffout}{A character string specify pdf output file.}
}
\description{
Plot BIC values over cluster size.
}
