% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\name{bplot}
\alias{bplot}
\title{bplot}
\usage{
bplot(obj, title, pdffout, maxPcnt = 0.8, ylab = expression(paste(log[2],
  "(TPM)")), isLog = FALSE, small = 0.05)
}
\description{
Boxplot of tpm.value slot of tpm object.
}
