% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\docType{methods}
\name{corplot}
\alias{corplot}
\alias{corplot,tpm,character-method}
\title{corplot}
\usage{
corplot(obj, pdffout)

\S4method{corplot}{tpm,character}(obj, pdffout)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{pdffout}{A character string specify pdf output file.}
}
\description{
Scatter plot of pairwise sample comparisons
}
