% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMaccess.R
\name{diffHeatmap}
\alias{diffHeatmap}
\title{diffHeatmap}
\usage{
diffHeatmap(tpm.value, col.idx, row.idx, pngfout, cutreek = NULL,
  cut.alg = c("pam", "hclust", "emmix"), rank.man = FALSE,
  log.it.already = FALSE, scale.it = TRUE, cluster_columns_par = TRUE,
  cluster_rows_par = TRUE, show_row_dend_par = FALSE, small = 0.05, ...)
}
\description{
heatmap during or after differential analysis
}
