% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\name{distplot}
\alias{distplot}
\title{distplot}
\usage{
distplot(obj, ylab, pdffout, probs = 0.85)
}
\description{
tpm.value either normalised to spikein or not
tpm.value either spikein, non-spikein, or all
idx for selecting differnetially expressed genes
}
