% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\name{heatcorplot}
\alias{heatcorplot}
\title{heatcorplot}
\usage{
heatcorplot(obj, pdffout)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{pdffout}{A character string specify pdf output file.}
}
\description{
Correlation coefficient heatmap of samples
}
