% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\docType{methods}
\name{hireplot}
\alias{hireplot}
\alias{hireplot,tpm-method}
\title{hireplot}
\usage{
hireplot(obj, pdffout)

\S4method{hireplot}{tpm}(obj, pdffout)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{pdffout}{A character string specify pdf output file.}
}
\description{
Hierarchical clustering of sample correlation coefficients
}
