% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMaccess.R
\name{limmaDiff}
\alias{limmaDiff}
\title{tpm.value either normalised to spikein or not
filtering of genes must be done before hand
export limmaDiff}
\usage{
limmaDiff(tpm.value.ori, grps, dout, pat, MA.it = TRUE, HEAT.it = TRUE,
  GO.it = TRUE, DiffOut.it = TRUE, logFCthresh = 1, PValtrhesh = 0.05,
  log2.it = T, small = 0.05)
}
\description{
tpm.value either normalised to spikein or not
filtering of genes must be done before hand
export limmaDiff
}
