% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\name{lineGraph}
\alias{lineGraph}
\title{lineGraph}
\usage{
lineGraph(obj, col.idx, row.idx, clusters, pdffout, log.it.already = FALSE,
  mean.it = FALSE, small = 0.05)
}
\description{
Line graphs of clustering
}
