% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\docType{methods}
\name{lmAdjCovar}
\alias{lmAdjCovar}
\alias{lmAdjCovar,numeric,data.frame,logical-method}
\title{lmAdjCovar}
\usage{
lmAdjCovar(x, covar, add.mean = TRUE)

\S4method{lmAdjCovar}{numeric,data.frame,logical}(x, covar, add.mean = TRUE)
}
\arguments{
\item{x}{A matrix of gene expression data with genes in the rows and samples in the cols.}

\item{covar}{A matrix/data.frame or vector of covariates.}

\item{add.mean}{A logical vector indicating whether to add group mean to the residuals}
}
\description{
Linear regression to adjust for batch effects.
}
