% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\docType{methods}
\name{loessnorm}
\alias{loessnorm}
\alias{loessnorm,tpm-method}
\title{loessnorm}
\usage{
loessnorm(obj, small = 0.05)

\S4method{loessnorm}{tpm}(obj, small = 0.05)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{small}{A numeric vector indicating the adjustment to the TPM values before log2 transformation.}
}
\description{
Loess normalization of gene or transcript expressions to expressions of spike-in.
}
