% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\docType{methods}
\name{MAchart}
\alias{MAchart}
\alias{MAchart,data.frame,character-method}
\title{MAchart}
\usage{
MAchart(dd, pdffout)

\S4method{MAchart}{data.frame,character}(dd, pdffout)
}
\arguments{
\item{dd}{A \code{data.frame} object, with columns of 'P.Value', 'logFC', 'nlogpval' and 'DEG'.}

\item{pdffout}{A character string specify pdf output file.}
}
\description{
plot of log2FC over pvalue during differential analysis.
}
