% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMaccess.R
\name{MAplot}
\alias{MAplot}
\title{plot of log2FC over pvalue during differential analysis
export MAplot}
\usage{
MAplot(dd, pngfout)
}
\description{
plot of log2FC over pvalue during differential analysis
export MAplot
}
