% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\docType{methods}
\name{rdcntnorm}
\alias{rdcntnorm}
\alias{rdcntnorm,tpm,data.frame-method}
\title{rdcntnorm}
\usage{
rdcntnorm(obj, stats)

\S4method{rdcntnorm}{tpm,data.frame}(obj, stats)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{stats}{A data.frame with named columns of 'spikein' and 'mm10' read counts.}
}
\description{
Normalization of gene or transcript expressions to expressions of spike-in based upon read counts in a table.
}
