% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\docType{methods}
\name{rmNonVar}
\alias{rmNonVar}
\alias{rmNonVar,tpm,numeric-method}
\title{rmNonVar}
\usage{
rmNonVar(obj, probs = 0.1)

\S4method{rmNonVar}{tpm,numeric}(obj, probs = 0.1)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{probs}{A probability threshold for removing invariably expressed genes.}
}
\description{
Filter genes or transcripts with expression variabilities above a given quantile.
}
