% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\name{sepSpike}
\alias{sepSpike}
\title{sepSpike}
\usage{
sepSpike(obj, invert = FALSE)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{invert}{A \code{logical} object, indicating whether to keep spike-in only (TRUE) or genes only (FALSE).}
}
\description{
Separate out spikein from genes based upon rownames of tpm.value slot of tpm object.
}
