% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMaccess.R
\name{SepTPMCnt}
\alias{SepTPMCnt}
\title{read in Salmon differential expression file}
\usage{
SepTPMCnt(fin)
}
\description{
Read in TPM from Salmon output and differential analysis
}
\author{
Erica Liu \email{ericaenjoy3@gmail.com}
}
