% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMaccess.R
\name{SepTPMCnt}
\alias{SepTPMCnt}
\title{Read in TPM from Salmon output and differential analysis}
\usage{
SepTPMCnt(fin)
}
\description{
Read in TPM from Salmon output and differential analysis
}
\author{
Erica Liu \email{ericaenjoy3@gmail.com}
import data.table
import limma
import affy
import ggplot2
import scatterplot3d
import RColorBrewer
import grDevices
import scales
import GGally
import GOtest
import ape
import pheatmap
import ComplexHeatmap
import circlize
import multiplot
import addgrids3d
export SepTPMCnt
}
