% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\name{diffHeatmap}
\alias{diffHeatmap}
\title{diffHeatmap}
\usage{
diffHeatmap(tpm.value, col.idx, row.idx, pdffout, cutreek = NULL,
  cut.alg = NULL, rank.man = FALSE, log.it.already = FALSE,
  scale.it = TRUE, cluster_columns_par = TRUE, cluster_rows_par = TRUE,
  show_column_dend_par = TRUE, show_row_dend_par = FALSE, small = 0.05,
  ...)
}
\description{
heatmap during or after differential analysis
}
