% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMaccess.R
\name{distplot}
\alias{distplot}
\title{distplot}
\usage{
distplot(tpm.value, grps, ylab, pngfout)
}
\description{
tpm.value either normalised to spikein or not
tpm.value either spikein, non-spikein, or all
idx for selecting differnetially expressed genes
}
