% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\docType{methods}
\name{distplot}
\alias{distplot}
\alias{distplot,tpm,character,character,numeric-method}
\title{distplot}
\usage{
distplot(obj, ylab, pdffout, probs = 0.85)

\S4method{distplot}{tpm,character,character,numeric}(obj, ylab, pdffout,
  probs = 0.85)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{ylab}{A character string specify y-labl on the plot.}

\item{pdffout}{A character string specify pdf output file.}

\item{probs}{A probability specifying the top quantile to use a ceilling in plotting.}
}
\description{
tpm.value either normalised to spikein or not
tpm.value either spikein, non-spikein, or all
idx for selecting differnetially expressed genes.
}
