% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\name{limmaDiff}
\alias{limmaDiff}
\title{limmaDiff}
\usage{
limmaDiff(obj, dout, pat, MA.it = TRUE, HEAT.it = TRUE, GO.it = TRUE,
  DiffOut.it = TRUE, logFCthresh = 1, PValtrhesh = 0.05, log2.it = TRUE,
  small = 0.05)
}
\description{
tpm.value either normalised to spikein or not
filtering of genes must be done before hand
}
