% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\docType{methods}
\name{PCAplot}
\alias{PCAplot}
\alias{PCAplot,tpm-method}
\title{PCAplot}
\usage{
PCAplot(obj, pdffout, fout = NULL, excl.col = NULL, ntop = Inf,
  isLog = FALSE, pca2d = FALSE, small = 0.05)

\S4method{PCAplot}{tpm}(obj, pdffout, fout = NULL, excl.col = NULL,
  ntop = Inf, isLog = FALSE, pca2d = FALSE, small = 0.05)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{pdffout}{A character string specify pdf output file.}

\item{fout}{A text output file of top N most variably expressed genes.}

\item{excl.col}{A numeric vector indicating columns to be excluded from the \code{tpm} slot of a \code{tpm} object.}

\item{ntop}{A numeric value indicating selection of the top N most variably expressed genes, or Inf (default).}

\item{isLog}{A logical value of whether the \code{tpm} data is already log2 transformed.}

\item{pca2d}{A logical value indicating whether 2D instead of 3D PCA.}

\item{small}{A numeric vector indicating the adjustment to the TPM values before log2 transformation.}
}
\description{
PCA plot (PC1, PC2 and PC3) of samples
}
