% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\name{rdcntnorm}
\alias{rdcntnorm}
\title{rdcntnorm}
\usage{
rdcntnorm(obj, stats)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{stats}{A data.frame with named columns of 'spikein' and 'mm10' read counts.}
}
\description{
Normalization of gene or transcript expressions to expressions of spike-in based upon read counts in a table.
}
