% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAMethods.R
\name{rmLow}
\alias{rmLow}
\title{rmLow}
\usage{
rmLow(obj, thresh)
}
\arguments{
\item{obj}{A \code{tpm} object.}

\item{thresh}{A TPM threshold.}
}
\description{
Filter of genes or transcripts with expression levels above given threshold in one or more samples.
}
