% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAConst.R
\name{SepTPMCnt}
\alias{SepTPMCnt}
\title{read in Salmon differential expression file}
\usage{
SepTPMCnt(fin)
}
\arguments{
\item{fin}{An input file name akin to "TPM_rd_merge.txt".}
}
\description{
Read in data file of tpm and count, and return tpm, cnt, rpm, tpm.grp.
}
